+++ application/tests/_ci_phpunit_test/replacing/filelist.sh
@@ -4,6 +4,7 @@ list="
 core/CodeIgniter.php
 core/Common.php
 core/Input.php
+core/Output.php
 core/Loader.php
 libraries/Upload.php
 helpers/download_helper.php